Fix native attention scratch storage and backward selection
Select native backward through the kernel config for BF16 D128 GQA ratios 4 or 8 on SM100. Other shapes retain the segmented port, and backward tile sweeps explicitly select that port. Keep the one-CTA schedule used in the throughput measurements.

Stop allocating the preprocess dQ accumulator whose output is discarded; the main backward kernel initializes its own accumulator. Share sparse metadata across heads with zero-stride CuTe views instead of materializing a copy per head. Use the K tile for dK/dV postprocessing so non-square tiles decode the correct layout, and widen zero-fill indices to 64 bits before multiplication to prevent overflow above 2^31 elements.

Give the segmented postprocess global-memory accumulator views instead of changing the upstream copy operation process-wide. This keeps native postprocessing independent of which kernel compiled first.

+++ lib/levanter/src/levanter/grug/attention/_fa4_thd.py
@@ -488,7 +488,7 @@ def _upstream_fa4_thd_backward_launcher(
             tidx, _, _ = cute.arch.thread_idx()
             bidx, _, _ = cute.arch.block_idx()
             flat = cute.make_tensor(tensor.iterator, cute.make_layout(cute.size(tensor)))
-            idx = bidx * self._num_threads + tidx
+            idx = cutlass.Int64(bidx) * self._num_threads + cutlass.Int64(tidx)
             if idx < cute.size(flat):
                 flat[idx] = cutlass.Float32(0.0)
